datas and templating optimization
