--- src/Form/Employee1Type.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Employee;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class Employee1Type extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('birthDate')
-            ->add('firstName')
-            ->add('lastName')
-            ->add('gender')
-            ->add('email')
-            ->add('photo')
-            ->add('hireDate')
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => Employee::class,
-        ]);
-    }
-}
