+++ src/Form/EmployeeType.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Employee;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Form\Extension\Core\Type\DateType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+
+class EmployeeType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('birthDate', DateType::class,[
+                'widget' => 'single_text',
+                'attr' => ['class' => 'js-datepicker'],
+            ])
+            ->add('firstName')
+            ->add('lastName')
+            ->add('gender', ChoiceType::class, [
+                'choices'  => [
+                    'Male' => 'M',
+                    'Female' => 'F',
+                    'Other' => 'X',
+                ],
+            ])
+            ->add('email', EmailType::class)
+            ->add('photo')
+            ->add('hireDate', DateType::class,[
+                'widget' => 'single_text',
+                'attr' => ['class' => 'js-datepicker'],
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Employee::class,
+        ]);
+    }
+}
