+++ templates/employee/new.html.twig
@@ -3,9 +3,11 @@
 {% block title %}New Employee{% endblock %}
 
 {% block body %}
+<div class="container">
     <h1>Create new Employee</h1>
 
     {{ include('employee/_form.html.twig') }}
 
     <a href="{{ path('app_employee_index') }}" class="btn btn-primary">back to list</a>
+</div>
 {% endblock %}
